requirements.txt
imapclient
pypdf
pdfminer.six
ocrmypdf
pandas
openpyxl
APScheduler
pydantic-settings
keyring
streamlit
watchdog
validate-docbr
loguru
python-dotenv
rapidfuzz
msal==1.31.0
SQLAlchemy
boto3
streamlit-authenticator
openai
plotly
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
python-multipart
